0.1
0.1

+++ includes/controle_excluir_cartao.php
@@ -1,6 +1,11 @@
 <?php
 require_once 'modelo_cartoes.php';
 
+if ($_SESSION['tipo_usuario'] !== 'admin') {
+    http_response_code(403); echo "<p>Acesso restrito ao administrador.</p>";
+    include '../includes/rodape.php'; exit;
+}
+
 if (isset($_GET['id'])) {
     $id_cartao = intval($_GET['id']);
     if (excluirCartao($id_cartao)) {
